+++ dolibarr_generators/generate_category.py
@@ -1,0 +1,27 @@
+import random
+import string
+import requests
+import base64
+import datetime
+
+
+from dolibarr_api import *
+from dolibarr_generators.generate_utils import *
+
+def generate_category(type):
+    # on boucle sur les lignes
+    url = urlBase + "categories"
+    data = {
+        "label": fake.company(),
+        "description": fake.catch_phrase(),
+        "type": type,
+        "status": 1, # actif
+    }
+    r = requests.post(url, headers=headers, json=data)
+    if r.status_code != 200:
+        print('Erreur lors de la création de la catégorie', r.status_code)
+        print (r.text)
+        return None
+
+    return 1
+
