+++ dolibarr_generators/generate_customer.py
@@ -1,0 +1,94 @@
+from faker import Faker
+import random
+import string
+import requests
+import base64
+import datetime
+
+
+from dolibarr_api import *
+from dolibarr_generators.generate_utils import *
+
+
+def generate_customer(dateCreate,enabledModule):
+    # on boucle sur les lignes
+    url = urlBase + "thirdparties"
+    typeTiers = random.choice([0, 1, 2])
+    typeFourn = 0
+    if createSupplier == 1:
+        typeFourn = random.choice([0, 1])
+
+    address, zip, town = get_random_address()
+
+    data = {
+        "name": fake.company(),
+        "address": address,
+        "zip": zip,
+        "town": town,
+        "phone": fake.phone_number(),
+        "email": fake.email(),
+        # "contact name": df['contact name'][index],
+        # "emailcontact": df['emailcontact'][index],
+        "client": typeTiers,
+        "fournisseur": typeFourn,
+        "country_id": random.randint(1, nbCountry),
+        "date_creation": dateCreate.strftime('%Y-%m-%d'),
+        # "useraffected": df['useraffected'][index],
+        # "dateupdate": df['dateupdate'][index],
+        # "proprietaire": df['proprietaire'][index],
+    }
+    r = requests.post(url, headers=headers, json=data)
+    if r.status_code != 200:
+        print('Erreur lors de la création du tiers', r.status_code)
+        print (r.text)
+        return None
+    else:
+        idSoc= r.text
+
+    # ajout de contact
+    url = urlBase + "thirdparties/" + idSoc + "/representative/"
+    for i in range(random.randint(0, 2)):
+        # on rajoute un utilisateur référent 
+        userRandom = get_random_user(retDataUser)
+        data = { }
+        r = requests.post(url + userRandom['id'], headers=headers, json=data)
+
+    
+    # ajout de contact
+    url = urlBase + "contacts/"
+    for i in range(random.randint(0, 3)):
+        # on rajoute des contacts externes
+        address, zip, town = get_random_address()
+
+        data = {
+            "lastname" : fake.last_name(),
+            "firstname" : fake.first_name(),
+            "socid" : idSoc,
+            "address": address,
+            "email": fake.email(),
+            "zip": zip,
+            "town": town,
+            "phone": fake.phone_number(),
+
+            "country_id": 1,
+        }
+        r = requests.post(url, headers=headers, json=data)
+        idContact = r.text 
+        # gestion des catégories de contact
+        if newCategorySocpeople > 0 and 'categorie' in enabledModule:
+            for i in range(random.randint(0, newCategorySocpeople)):
+                # on rajoute une catégorie aléatoire
+                url = urlBase + "categories/" + str(random.choice(retDataCategContact)['id']) + "/objects/contact/" + str(idContact)
+                data = { }
+                r = requests.post(url, headers=headers, json=data)
+
+    # gestion des catégories de tiers
+    if newCategoryCustomer > 0 and 'categorie' in enabledModule:
+        for i in range(random.randint(0, newCategoryCustomer)):
+            # on rajoute une catégorie aléatoire
+            #categories/5/objects/product/100
+            url = urlBase + "categories/" + str(random.choice(retDataCategCustomer)['id']) + "/objects/customer/" + str(idSoc)
+            data = { }
+            r = requests.post(url, headers=headers, json=data)
+
+    return 1
