+++ dolibarr_generators/generate_opportunity.py
@@ -1,0 +1,3 @@
+
+def generate_opportunity ():
+    print("à implementer")
