+++ dolibarr_generators/generate_utils.py
@@ -1,0 +1,14 @@
+from faker import Faker
+from datetime import datetime
+
+# Choisi la langue pour les fausses données de faker.
+fake = Faker('fr_FR')
+
+# récupération de l'année en cours
+yearNow = datetime.now().year
+
+def get_random_address():
+    fulladdress = fake.address()
+    arrayaddress = fulladdress.split("\n")
+    arraycpville = arrayaddress[1].split(" ")
+    return arrayaddress[0], arraycpville[0], arraycpville[1]
